+++ src/router/index.ts
@@ -1,6 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router';
 import { ROUTES } from '@/router/routes';
-import HomePage from '@/pages/HomePage/HomePage.vue';
+import HomePage from '@/modules/home/pages/home-page.vue';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
